fix warnings

+++ lib/dat/Input.cpp
@@ -15,7 +15,7 @@ nlohmann::json Input::initKeyMap(){
 bool Input::isDown(const std::string actionName){
     bool down = false;
     std::vector<SDL_Keycode> keycodes = getAction(actionName);
-    for (int i = 0; i < keycodes.size(); i++){
+    for (size_t i = 0; i < keycodes.size(); i++){
         if (keystates[SDL_GetScancodeFromKey(keycodes[i])]){
             down = true;
             break;
@@ -27,7 +27,7 @@ bool Input::isDown(const std::string actionName){
 std::vector<SDL_Keycode> Input::getAction(const std::string actionName) {
     nlohmann::json keys = keyMap["actions"][actionName];
     std::vector<SDL_Keycode> keycodes;
-    for (int i=0; i<keys.size();i++){
+    for (size_t i=0; i<keys.size();i++){
         std::string keyName = keys[i].get<std::string>();
         keycodes.push_back(SDL_GetKeyFromName(keyName.c_str()));
     }
